requirements.txt
flask==3.1.0
requests==2.31.0
google-genai